sunspot_prediction/requirements.txt
xgboost
pandas
scikit-sklearn